added the half crate license
Arcane Fortune doesn't presently use functions from that crate -- so I suppose I hadn't added it into this file which was originally for the binary distributions of the game. however, the ironnetpro crate contains some of it for some types of gpu training, so it ought to be included here too even if the game never calls that code.